+++ lib/grape/validations/regexp.rb
@@ -2,7 +2,8 @@ module Grape
   module Validations
     class RegexpValidator < SingleOptionValidator
       def validate_param!(attr_name, params)
-        if params[attr_name] && !(params[attr_name].to_s =~ @option)
+        if params.has_key?(attr_name) &&
+          (params[attr_name].nil? || !(params[attr_name].to_s =~ @option))
           raise Grape::Exceptions::Validation, param: @scope.full_name(attr_name), message_key: :regexp
         end
       end